+++ modules/graphics/functions/cla.m
@@ -26,10 +26,12 @@ function varargout = cla(varargin)
   ax.ZScale = 'linear';
   ax.XTickMode = 'auto';
   ax.YTickMode = 'auto';
-  ax.ZtickMode = 'auto';
+  ax.ZTickMode = 'auto';
   ax.XTickLabelMode = 'auto';
   ax.YTickLabelMode = 'auto';
   ax.ZTickLabelMode = 'auto';
+  ax.ColorOrderIndex = 1;
+  ax.LineStyleOrderIndex = 1;
   if nargout == 1
     varargout{1} = ax;
   else