+++ modules/graphics/tests/test_copper.m
@@ -12,6 +12,6 @@
 f = figure();
 surf(peaks);
 colormap('copper');
-r = f.ColorMap;
+r = f.Colormap;
 assert_isapprox(r(1, :), [0, 0, 0], 1e-3); 
 %=============================================================================